pyproject.toml
[tool.poetry]
name = "open-anime-tracker"
version = "0.1.0"
description = ""
authors = ["FehrIsFair <fehrisfair@proton.me>"]
readme = "README.md"
packages = [{include = "open_anime_tracker", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
alembic = "^1.13.1"
sqlalchemy = "^2.0.25"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
uuid = "^1.30"
flask-classful = "^0.16.0"
flask-bcrypt = "^1.0.1"
python-dotenv = "^1.0.1"
flask-cors = "^4.0.0"
flask-restful = "^0.3.10"
requests = "^2.31.0"
flask-login = "^0.6.3"
flask-security = "^3.0.0"
flask-session = "^0.8.0"
redis = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
